+++ src/styles/themes/index.js
@@ -1,0 +1,2 @@
+export { default as future } from './future'
+export { default as legacy } from './legacy'